+++ src/redis.error.ts
@@ -4,9 +4,10 @@ export class RedisCoreError extends CoreError {
   name: string
 
   constructor(message?: string) {
-    super(message);
-    this.name = "RedisCoreError";
-    this.message = message;
-    this.stack = (new Error(this.message)).stack;  //`${this.message}\n${new Error().stack}`;
+    super();
+    Object.setPrototypeOf(this, RedisCoreError.prototype);
+    this.name = this.constructor.name
+    this.message = message
+    if (Error.captureStackTrace) Error.captureStackTrace(this, RedisCoreError);
   }
 }