+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.0] 2025-04-07
+### Updated
+* Update `tydet-core` to `2.0.0`.
+* Handle the new events for `restart` and `eject` service methods.
+### Added
+* Add `onReady()` callback.
+
 ## [v1.0.4] 2024-10-08
 ### Fix
 * Downgrade chalk.